# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "For the administration of analytics flows.",
        "title": "Analytics-Flow-Repo-V2 API",
        "contact": {},
        "license": {
            "name": "Apache-2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "{version}"
    },
    "basePath": "/",
    "paths": {
        "/flow": {
            "get": {
                "description": "Gets all flows",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Flow"
                ],
                "summary": "Get flows",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/lib.FlowsResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/flow/": {
            "put": {
                "description": "Validates and stores a flow",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Flow"
                ],
                "summary": "Create flow",
                "parameters": [
                    {
                        "description": "Create flow",
                        "name": "flow",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/lib.Flow"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/lib.FlowCreateResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "424": {
                        "description": "Failed Dependency",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/flow/{id}": {
            "get": {
                "description": "Gets a single flow",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Flow"
                ],
                "summary": "Get flow",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Flow ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/lib.Flow"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/flow/{id}/": {
            "post": {
                "description": "Validates and updates a flow",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "Flow"
                ],
                "summary": "Update flow",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Flow ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Update flow",
                        "name": "flow",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/lib.Flow"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "424": {
                        "description": "Failed Dependency",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "delete": {
                "description": "Deletes a flow. Refused with 409 while pipelines or smart services use it, unless an administrator passes force. Answers 424 when the pipeline registry cannot be asked and 502 when the smart-service-repository cannot; nothing is deleted then, also with force.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Flow"
                ],
                "summary": "Delete flow",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Flow ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "boolean",
                        "default": false,
                        "description": "Administrators only: delete although pipelines or smart services still use the flow",
                        "name": "force",
                        "in": "query"
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "409": {
                        "description": "Conflict",
                        "schema": {
                            "$ref": "#/definitions/lib.StillInUseResponse"
                        }
                    },
                    "424": {
                        "description": "Failed Dependency",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "502": {
                        "description": "Bad Gateway",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/info": {
            "get": {
                "description": "Get basic service and runtime information.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Info"
                ],
                "summary": "Get service info",
                "responses": {
                    "200": {
                        "description": "info",
                        "schema": {
                            "$ref": "#/definitions/srv_info_hdl.ServiceInfo"
                        }
                    },
                    "500": {
                        "description": "error message",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/operators/{id}/usage": {
            "get": {
                "description": "Counts the flows of all users that contain a node of the operator and lists the ones the caller may read.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Operator"
                ],
                "summary": "Get usage of an operator",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Operator ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/lib.OperatorFlowUsage"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "lib.Cell": {
            "type": "object",
            "properties": {
                "config": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/lib.ConfigValue"
                    }
                },
                "cost": {
                    "type": "integer"
                },
                "deploymentType": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "image": {
                    "type": "string"
                },
                "inPorts": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "name": {
                    "type": "string"
                },
                "operatorId": {
                    "type": "string"
                },
                "outPorts": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "position": {
                    "$ref": "#/definitions/lib.CellPosition"
                },
                "source": {
                    "$ref": "#/definitions/lib.CellLink"
                },
                "target": {
                    "$ref": "#/definitions/lib.CellLink"
                },
                "type": {
                    "type": "string"
                },
                "version": {
                    "type": "integer"
                }
            }
        },
        "lib.CellLink": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string"
                },
                "magnet": {
                    "type": "string"
                },
                "port": {
                    "type": "string"
                }
            }
        },
        "lib.CellPosition": {
            "type": "object",
            "properties": {
                "x": {
                    "type": "number"
                },
                "y": {
                    "type": "number"
                }
            }
        },
        "lib.ConfigValue": {
            "type": "object",
            "properties": {
                "name": {
                    "type": "string"
                },
                "type": {
                    "type": "string"
                }
            }
        },
        "lib.Flow": {
            "type": "object",
            "properties": {
                "_id": {
                    "type": "string"
                },
                "dateCreated": {
                    "type": "string"
                },
                "dateUpdated": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "image": {
                    "type": "string"
                },
                "model": {
                    "$ref": "#/definitions/lib.Model"
                },
                "name": {
                    "type": "string"
                },
                "userId": {
                    "type": "string"
                }
            }
        },
        "lib.FlowCreateResponse": {
            "type": "object",
            "properties": {
                "_id": {
                    "type": "string"
                }
            }
        },
        "lib.FlowRef": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                }
            }
        },
        "lib.FlowsResponse": {
            "type": "object",
            "properties": {
                "flows": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/lib.Flow"
                    }
                },
                "total": {
                    "type": "integer"
                }
            }
        },
        "lib.Model": {
            "type": "object",
            "properties": {
                "cells": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/lib.Cell"
                    }
                }
            }
        },
        "lib.OperatorFlowUsage": {
            "type": "object",
            "properties": {
                "flows": {
                    "type": "integer"
                },
                "readable": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/lib.FlowRef"
                    }
                }
            }
        },
        "lib.SmartServiceRef": {
            "type": "object",
            "properties": {
                "design_id": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                }
            }
        },
        "lib.StillInUseResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                },
                "instances": {
                    "type": "integer"
                },
                "pipelines": {
                    "type": "integer"
                },
                "readable": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/lib.SmartServiceRef"
                    }
                },
                "releases": {
                    "type": "integer"
                }
            }
        },
        "srv_info_hdl.ServiceInfo": {
            "type": "object",
            "properties": {
                "mem_alloc": {
                    "type": "string"
                },
                "mem_alloc_b": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "up_time": {
                    "type": "string"
                },
                "up_time_ns": {
                    "type": "integer"
                },
                "version": {
                    "type": "string"
                }
            }
        }
    }
}