# lib/models.go
/*
 * Copyright 2025 InfAI (CC SES)
 *
 * Licensed under the Apache License, Version 2.0 (the "License");
 * you may not use this file except in compliance with the License.
 * You may obtain a copy of the License at
 *
 *      http://www.apache.org/licenses/LICENSE-2.0
 *
 * Unless required by applicable law or agreed to in writing, software
 * distributed under the License is distributed on an "AS IS" BASIS,
 * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 * See the License for the specific language governing permissions and
 * limitations under the License.
 */

package lib

import (
	"time"

	"go.mongodb.org/mongo-driver/bson/primitive"
)

type FlowsResponse struct {
	Flows []Flow `json:"flows"`
	Total int64  `json:"total"`
}
type Flow struct {
	Id          *primitive.ObjectID `bson:"_id,omitempty" json:"_id,omitempty"`
	Name        string              `json:"name,omitempty"`
	Description *string             `json:"description,omitempty"`
	Model       Model               `json:"model,omitempty"`
	Image       *string             `json:"image,omitempty"`
	UserId      string              `bson:"userId,omitempty" json:"userId,omitempty"`
	DateCreated time.Time           `bson:"dateCreated,omitempty" json:"dateCreated,omitempty"`
	DateUpdated time.Time           `bson:"dateUpdated,omitempty" json:"dateUpdated,omitempty"`
}

type FlowCreateResponse struct {
	Id string `json:"_id"`
}

type Model struct {
	Cells []Cell `json:"cells,omitempty"`
}

type Cell struct {
	Type           string         `json:"type,omitempty"`
	InPorts        []string       `json:"inPorts,omitempty"`
	OutPorts       []string       `json:"outPorts,omitempty"`
	Name           *string        `json:"name,omitempty"`
	Image          *string        `json:"image,omitempty"`
	OperatorId     *string        `json:"operatorId,omitempty"`
	Position       *CellPosition  `json:"position,omitempty"`
	Source         *CellLink      `json:"source,omitempty"`
	Target         *CellLink      `json:"target,omitempty"`
	Id             string         `json:"id,omitempty"`
	Config         *[]ConfigValue `json:"config,omitempty"`
	Cost           *int64         `json:"cost,omitempty"`
	DeploymentType *string        `json:"deploymentType,omitempty"`
	Version        *int           `json:"version,omitempty"`
}

type CellPosition struct {
	X float64 `json:"x"`
	Y float64 `json:"y"`
}

type CellLink struct {
	Id     string `json:"id"`
	Magnet string `json:"magnet"`
	Port   string `json:"port"`
}

type ConfigValue struct {
	Name string `json:"name,omitempty"`
	Type string `json:"type,omitempty"`
}

type OperatorFlowCount struct {
	OperatorID string      `bson:"_id"`
	Flows      []FlowCount `bson:"flows"`
}

type FlowCount struct {
	FlowID *primitive.ObjectID `bson:"flowId"`
	Count  int32               `bson:"count"`
}

// OperatorFlowUsage counts the flows of all users that contain an operator; Readable lists only those the caller may read.
type OperatorFlowUsage struct {
	Flows    int       `json:"flows"`
	Readable []FlowRef `json:"readable"`
}

type FlowRef struct {
	Id   string `json:"id"`
	Name string `json:"name"`
}

// SmartServiceRef names a smart service design that uses a flow and that the caller may read.
type SmartServiceRef struct {
	Id       string `json:"id"`
	DesignId string `json:"design_id"`
	Name     string `json:"name"`
}

// SmartServiceUsage is the smart-service-repository's answer: Releases and Instances count over
// all users, Readable lists only what the caller may read.
type SmartServiceUsage struct {
	Releases  int               `json:"releases"`
	Instances int               `json:"instances"`
	Readable  []SmartServiceRef `json:"readable"`
}

// StillInUseResponse is the 409 body of the flow delete: Pipelines counts the pipelines using the
// flow, the other fields are the smart-service-repository's answer.
type StillInUseResponse struct {
	Error     string            `json:"error"`
	Pipelines int               `json:"pipelines"`
	Releases  int               `json:"releases"`
	Instances int               `json:"instances"`
	Readable  []SmartServiceRef `json:"readable"`
}

// DeleteOptions carries what a flow delete needs besides the ids. Force deletes a flow that
// pipelines or smart services still use; the caller has to have checked that the user may do so.
type DeleteOptions struct {
	Force bool
}